% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.stepwise.FDR.R
\name{forward.stepwise.FDR}
\alias{forward.stepwise.FDR}
\title{This is an R function to perform stepwise regression based on a "nested model" F test for inclusion/exclusion
of a predictor.  To keep it simple, I made no provision for forcing certain variables to be included in
all models, did not allow for specification of a data frame, and skipped some consistency checks (such as whether
the initial model is a subset of the full model).}
\usage{
forward.stepwise.FDR(full.model, initial.model, XX)
}
\arguments{
\item{full.model}{A model formula}

\item{initial.model}{A model formula}

\item{XX}{a data.frame}
}
\description{
One other note: since the code uses R's drop1 and add1 functions, it respects hierarchy in models. That is,
regardless of p values, it will not attempt to drop a term while retaining a higher order interaction
involving that term, nor will it add an interaction term if the lower order components are not all present.
(You can of course defeat this by putting interactions into new variables and feeding it what looks like
a first-order model.)
}
\details{
Consider this to be "beta" code (and feel free to improve it).  I've done very limited testing on it.

Author: Paul A. Rubin (rubin@msu.edu)
}
\examples{
full.model <- mpg~ .
initial.model <- mpg~ 1
aa<-forward.stepwise.FDR(full.model, initial.model,XX=mtcars)
}
\keyword{Bonferroni}
\keyword{linear}
\keyword{regression}
\keyword{stepwise}

