% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEPWISE.R
\name{STEPWISE}
\alias{STEPWISE}
\title{A variable selection and parameter estimation method}
\usage{
STEPWISE(x, y, family, eta1 = 0, eta2 = 1, initvars = NULL)
}
\arguments{
\item{x}{A set of predictor variables. Should be either a numeric matrix or a data frame.}

\item{y}{The outcome.}

\item{family}{A description of the error distribution and link function to be used in the model.
This can be a character string naming a family function, a family function or the result of a call to a family function.}

\item{eta1}{A stopping criteria for controlling the false negatives and positives at the forward step of the method.
Can be any non-negative number. By default, eta1 = 0.}

\item{eta2}{A stopping criteria for controlling the false negatives and positives at the backward step of the method.
Can be any non-negative number. By default, eta2 = 1.}

\item{initvars}{An initial set of predictors included in the model. Should be a character vector with variable names.
By default, initvars = NULL.}
}
\value{
A list of four elements: "finalset", "finalset2", "coef1" and "coef2". "Finalset" is a set of predictors selected
 by the forward step of the STEPWISE method; "Finalset2" is a set of predictors selected after completing both forward and
 backward steps of the STEPWISE method and should be considered as the final result of the method. "Coef1" and "Coef2"
 are parameter estimates of the GLM models built with "Finalset" and "Finalset2", respectively.
}
\description{
This function implements STEPWISE method that performs a variable selection and consistent estimation of
the model parameters in Generalized Linear Models (GLM).
}
